Backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
python-dotenv
requests
python-dateutil
python-telegram-bot
pytesseract
opencv-python
ultralytics
whisper
scipy
vaderSentiment